+++ apps/web/src/app/archive/wish/[id]/_hooks/usePatchWish.ts
@@ -3,30 +3,28 @@ import { useRouter } from 'next/navigation';
 import { toast } from 'sonner';
 
 import { ROUTES } from '@/consts/route';
-import { useBackWithFallback } from '@/hooks/useBackWithFallback';
-import type { PatchItemRequestT } from '@/types/item';
 import { getApiErrorStatus, isGlobalNetError } from '@/utils/apiError';
 import { getApiErrorMessage } from '@/utils/getApiErrorMessage';
 
 import { patchWish } from '../_apis/patchWish';
+import type { PatchWishRequestT } from '../_types/wish';
 
 export const usePatchWish = (wishId: number) => {
   const router = useRouter();
-  const backWithFallback = useBackWithFallback();
   const queryClient = useQueryClient();
 
   const { mutate: patchWishMutation, isPending: isPatchWishPending } = useMutation({
-    mutationFn: (body: PatchItemRequestT) => {
+    mutationFn: (body: PatchWishRequestT) => {
       const formData = new FormData();
       if (body.name) formData.append('name', body.name);
       if (body.price) formData.append('price', String(body.price));
       if (body.image) formData.append('image', body.image);
+      if (body.memo) formData.append('memo', body.memo);
       return patchWish(wishId, formData);
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['wish', wishId] });
       queryClient.invalidateQueries({ queryKey: ['wishlists'] });
-      backWithFallback(ROUTES.WISHLIST);
     },
     onError: error => {
       if (isGlobalNetError(error)) return;
