+++ apps/web/src/components/common/item-info-screen/ItemDetailView.tsx
@@ -5,17 +5,17 @@ import Spacing from '@/components/spacing';
 import ItemInfoCard from './ItemInfoCard';
 import ItemLinkBanner from './ItemLinkBanner';
 import ItemMemoCard from './ItemMemoCard';
-import type { PriceRefreshT, ReadyItemInfoT } from './itemInfoScreen.const';
+import type { MemoT, PriceRefreshT, ReadyItemInfoT } from './itemInfoScreen.type';
 
 type ItemDetailViewProps = {
   item: ReadyItemInfoT;
-  hasMemo: boolean;
+  memo?: MemoT;
   priceRefresh?: PriceRefreshT;
   onEdit: () => void;
 };
 
 /** 정보를 정상적으로 가져온 상품의 조회 화면 */
-function ItemDetailView({ item, hasMemo, priceRefresh, onEdit }: ItemDetailViewProps) {
+function ItemDetailView({ item, memo, priceRefresh, onEdit }: ItemDetailViewProps) {
   return (
     <>
       <div className="relative mt-5 aspect-square w-full overflow-hidden rounded-xl bg-gray-50">
@@ -32,10 +32,10 @@ function ItemDetailView({ item, hasMemo, priceRefresh, onEdit }: ItemDetailViewP
         onEdit={onEdit}
       />
 
-      {hasMemo && (
+      {memo && (
         <>
           <Spacing size={12} />
-          <ItemMemoCard />
+          <ItemMemoCard memo={memo.value} onSave={memo.save} />
         </>
       )}
     </>
