+++ apps/web/src/components/common/item-info-screen/ItemMemoCard.tsx
@@ -9,13 +9,13 @@ import ItemMemoDialog from './ItemMemoDialog';
 
 const MEMO_PLACEHOLDER = '위시에 대한 생각을 메모해보세요. 메모는 나만 볼 수 있어요.';
 
-/**
- * 메모 카드 — 탭하면 메모 다이얼로그가 열린다
- *
- * TODO: 메모 조회·저장 API 가 아직 없어 화면 안에서만 유지된다. API 연동 시 상위에서 값을 주입할 것
- */
-function ItemMemoCard() {
-  const [memo, setMemo] = useState('');
+type ItemMemoCardProps = {
+  memo: string;
+  onSave: (memo: string) => void;
+};
+
+/** 메모 카드 — 탭하면 메모 다이얼로그가 열린다 */
+function ItemMemoCard({ memo, onSave }: ItemMemoCardProps) {
   const [isDialogOpen, setIsDialogOpen] = useState(false);
 
   return (
@@ -43,7 +43,8 @@ function ItemMemoCard() {
         open={isDialogOpen}
         onOpenChange={setIsDialogOpen}
         memo={memo}
-        onSave={setMemo}
+        /** 그대로 확인만 눌렀다면 저장하지 않는다 */
+        onSave={nextMemo => nextMemo !== memo && onSave(nextMemo)}
       />
     </>
   );
